+++ components/fxa-client/src/error.rs
@@ -5,11 +5,12 @@
 use error_support::{ErrorHandling, GetErrorHandling};
 use std::string;
 
+#[derive(uniffi::Error, Debug, thiserror::Error)]
+#[uniffi(flat_error)]
 /// Public error type thrown by many [`FirefoxAccount`] operations.
 ///
 /// Precise details of the error are hidden from consumers. The type of the error indicates how the
 /// calling code should respond.
-#[derive(Debug, thiserror::Error)]
 pub enum FxaError {
     /// Thrown when there was a problem with the authentication status of the account,
     /// such as an expired token. The application should [check its authorization status](