+++ components/support/rc_crypto/nss/nss_sys/build.rs
@@ -3,5 +3,8 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 fn main() {
+    #[cfg(feature = "mozbuild-rustlib")]
+    nss_build_common::link_nss_rustlib().unwrap();
+    #[cfg(not(feature = "mozbuild-rustlib"))]
     nss_build_common::link_nss().unwrap();
 }